add authorization

+++ routes/events.py
@@ -1,4 +1,6 @@
 # routes/events.py
+# TODO: 몇가지 함수에서 코드에러 발생으로 설정해둔 예외가 발생하지 않고 500 서버오류 발생.
+# 이유는 bool 값엔 .creator 등등이 없다. 고로 추후에 수정 바람.
 
 from beanie import PydanticObjectId
 from fastapi import APIRouter, HTTPException, status, Depends
@@ -13,7 +15,7 @@ event_router = APIRouter(
     tags=["Events"]
 )
 
-event_database = Database(Event)
+event_database = Database(Event)  # CRUD를 위한 Database 클래스의 인스턴스 생성
 
 # events = []  # DB서버 구축하면서 필요없어짐
 
@@ -61,23 +63,42 @@ async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
 
 @event_router.put("/{id}", response_model=Event)
 async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
-    update_event = await event_database.update(id, body)
-    if not update_event:
+    event = await event_database.get(id)  # DB에서 해당id 이벤트 Document를 불러온다.
+    if not event:  # 이벤트가 있는지 확인
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist"
         )
+    if event.creator != user:  # 현재 사용자가 해당 이벤트를 수정할 권한이 있는지 먼저 확인한다.
+        raise HTTPException(  # 권한이 없다면 HTTP_400_BAD_REQUEST 예외를 발생시킨다.
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
+    # 이 밑에서부터는 업데이트 작업 수행
+    update_event = await event_database.update(id, body)
+    # if not update_event:
+    #     raise HTTPException(
+    #         status_code=status.HTTP_404_NOT_FOUND,
+    #         detail="Event with supplied ID does not exist"
+    #     )
     return update_event
     
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate))-> dict:
-    event = await event_database.delete(id)
-    if not event:
+    event = await event_database.get(id)  # DB에서 해당id 이벤트 Document를 불러온다. 
+    if not event:  # 이벤트가 없다면 예외 발생
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist"
         )
+    if event.creator != user:  # 현재 사용자가 해당 이벤트를 수정할 권한이 있는지 먼저 확인한다.
+        raise HTTPException(  # 권한이 없다면 HTTP_404_NOT_FOUND 예외를 발생시킨다.
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Operation not allowed"
+        )
+    # 이 밑에서부터는 delete 작업 수행
+    event = await event_database.delete(id)
     return {
         "message": "Event deleted successfully."
     }
